requirements.txt
boto3==1.17.112
Pillow==10.0.0